Step 1 - Project Setup
Created project in Unreal Engine 5.2 with both C++ and blueprints enabled and imported all assets needed for the course. This is a top down style of Role Playing Game (RPG) featuring a character with wizard-like gameplay abilities. The purpose of this project is to learn all about the Gameplay Ability System (GAS) and how to implement it on both single and multiplayer modes.