Step 2 - Character Class for Player and Enemies
Created a BaseCharacter class to be used in both Enemy and Player Character classes. Used EnhancedInput features to move the Player and highlight Enemy class with mouse cursor. Created Animation blueprints for all Characters and established a GameMode to use for the game going forward. Used PostProcessing features to help with the highlighting of Enemies on the screen. This will assist with spells towards Enemies in future commits.