Step 4 - Attributes
Setup default attributes for Health and Mana and added appropriate Attribute Accessors. Created an Effect Actor class for pickup items that affect the Health, Mana (and eventually other Attributes) of objects when overlapped in the game. Next commit will focus on creating a Game User Interface to show the levels of Health and Mana on the screen.