Step 5 - RPG Game UI
Created widget classes as globes on an Overlay Widget to show current values for Health and Mana attributes on the character. Using broadcast and callbacks on these attributes from the widget controller, each widget can listen for changes which then get updated on the screen. An effect actor was created to test the changes of attribute amounts, but eventually we will be using gameplay effects to handle these changes.