Step 7 - Gameplay Tags
Created a few basic Gameplay Tags using three different methods including directly in the editor project settings, from Data Tables and using Gameplay Effects. Created a Widget Data Table which was used to identity items (Actors) picked up in the game. Used an animated widget to display on the screen during item pickup. Replaced all the callbacks in the Widget Controller with Lambdas. Created a ghost globe which follows the value of a vital attribute with a delay. Finally found a way to property clamp attributes in the PostGameplayEffectExecute function. In the next commit we will focus on setting up all the attributes that will be used in the game.