Step 9 - Attribute Menu
Created a widget that displays the Attribute Menu on the main HUD. Created a button on the main HUD to accomplish this task (while the menu widget is open, this button is disabled). Set the Attribute Tag on each item in the widget so that all primary and secondary attributes appear in the widget showing the correct name and value. The attribute menu also had a button that closes the widget and sends a signal to the button to re-enable. Also set up mapping tags in C++ to broadcast attribute data to the widget. The next commit will focus on Gameplay Abilities such as Click To Move and Minerva's first Spell!