Step 10 - Gameplay Abilities
Created a Gameplay Abilities Class to house Inputs for 6 Gameplay Abilities tied to 6 User Inputs for Left and Right Mouse Buttons, and numbers on the keyboard from 1 through 4. Setup a Click to Move feature that allows the Character to move by holding down the left mouse button or simply clicking the left mouse button for the character to auto run to the clicked location. Also implemented the Character's first Gameplay Ability (Firebolt Projectile Spell)  that shoots a fireball from the Character's Weapon when clicked on an Enemy!