Step 11 - Ability Tasks
Updated the Projectile to orient in the direction of the Enemy OR in the direction the player clicks (if Shift Key is held down). Added sounds to the projectile for casting the spell, flying across the screen and the impact of hitting the Enemy. Added Motion Warping so the Character turns in the direction she casts the Firebolt Spell. Added a Gameplay Effect to the Projectile to cause Damage to the Enemy when hit. Also added Enemy Health and Ghost progress bars to show the level of Health of Enemies when they are hit with the Projectile. Before setting up Damage, will briefly set up the Character RPG Classes in the next commit.