Step 13 - Damage
Created animations for Enemy reactions including Hit React and Death. Added a dissolve effect to the Enemy a few seconds after it dies. Created a floating text widget to show the number of damage points the Enemy takes when hit by the Player's weapon. Created a Execution Calculation class to test different damage effects on the Enemy.