Step 14 - Advanced Damage Techniques
Added a floating text that displays the amount of damage inflicted on an Enemy during impact of the Player's Weapon. Added a message and matching text color in the event of a Critical Hit or Block of the Enemy. Added Resistance Attributes to the Character Base Class and implemented boiler plate code that appears on the UI Attribute Menu. Testing operation of the game in Multiplayer to ensure there aren't any issues.