Step 15 - Enemy AI
Created a Blackboard and Behavior Tree for the Enemies in the game. Setup tree so that Enemies find the Player and move into range for both Melee and Ranged Enemy attacks. Ran a few EQS tests to make sure we get the behavior we are expecting from Enemies, especially when there is an obstacle to move around. Next commit will focus on the actual combat of Enemy melee attacks.