Step 17 - Enemy Ranged Attacks
Created a blueprint for a Rock Projectile. Created montages for both Ranged Attacked Goblin (on the Enemy) and Slingshot Attack (on the Weapon itself). Setup an animation blueprint for the Slingshot attack that allows natural movement of the Weapon during the ranged attack of the Enemy. This was the first time I've had to do animation of an actual Weapon and it worked out pretty well.