Step 18 - Enemy Spell Attacks
Created a Goblin Shaman Enemy that can hurl Fire Spells at the Player's Character. Since the infrastructure was already in place, no coding was needed to make the character. Simply created a new Enemy, marked it as an Elementalist, and adjusted the Ranged Gameplay Ability to shoot Fireballs in stead of Rocks like the Slingshot Goblin does. Then all I had to do was add a material that would dissolve when dead. Speaking of dead, I did have to add one line of code to the Enemy Character C++ class to set the IsDead variable on the Behavior Tree to stop Enemies from shooting once they died.