Step 19 - Enemy Finishing Touches
Added sounds for react, death and impact for Ghoul and all the Goblin enemies. Added a swipe effect to the hands for the Ghoul attack. Created two Demon minions (Ranger & Warrior) for the Shamon to summon. Shamon can now summon 5 Demons at a time using a cool particle effect. Completed dissolve material for all Enemies. Created an Elementalist behavior tree so that Shamon can still shoot fireballs when more then 2 Demons are alive. Finally, I added some juice to make the summoning of the Demons look better. Quite a bit of coding and game content for this commit!