Step 20 - Level Tweaks
Created a new Dungeon level with appropriate level lighting and Post Processing. Fixed an issue with the Texture Streaming Pool Over Budget error. Added a couple of Flame Pillars for that real dungeon effect. Added a series of Fade Actors that fade away as the camera following the actor blocks the visibility of the Character. Had to create a lot of new materials to pull this off. Next commit will focus on the Cost and Cooldown effects with relation to Mana and Spells.