Step 21 - Cost and Cooldown
Created a new widget for the Spell Globe. Added several of these Spell Globes to another new widget showing several Spell Globes along with the Health and Mana Globes. Also added an XP Bar to this widget (it won't be coded until the next commit). Created an Ability Info Data Asset table to broadcast information to the Spell Globes. Created Gameplay Effects for both Cost and Cooldown and applied these to Minerva's first (and currently only) Gameplay Ability - Firebolt. Created a Cooldown Async Task to show the updated cooldown time in the HUD after firing an ability. The next commit will focus on experience points or Minerva and her ability to level up.