Step 22 - Experience Points and Leveling Up
Created a Level Up Data Asset containing variables related to Experience Points (XP). Updated Player State and Character Attributes to handle Incoming XP. The Data Asset specifically identifies how much XP the Player needs to reach the next level (commonly known as Leveling Up). Added a variable to the Enemy Class for XP that is rewarded to the Player when the Enemy is killed. Created a delegate to Listen for XP Events that (one) knows how much XP to award to Player and (two) allows Player to Level Up at certain thresholds. Added a Niagara Effects component to the Character for Leveling Up. Finally, added a Widget to the overlay Screen that displays a level up message complete with animation, sound and Niagara Effects. This all came together quite nicely. Next commit will focus on awarded attribute points after leveling up and how to "spend" them on the Character.