Step 23 - Attribute Points
Created an Attribute Points member variable in the Player State and updated the Attribute Menu widget to show these points when the player levels up. Updated the buttons in the Attribute Menu to disburse points to appropriate Player Attribute buttons. Fixed an issue with filling the Health and Mana globes when leveling up. Did some minor changes to the background image of the Attribute Menu. Finally, added some sounds for OnClicked and OnHovered events to the various buttons in the game. The next commit will focus on creating and implementing the Spell Menu.